Rename DiscordOptionMapper to DiscordCommandParameterMapper, implement ParameterNamingStrategy
