+++ core/src/main/java/me/mastercapexd/auth/discord/command/DiscordCommandRegistry.java
@@ -3,19 +3,23 @@ package me.mastercapexd.auth.discord.command;
 import java.util.Collections;
 
 import com.bivashy.auth.api.AuthPlugin;
+import com.bivashy.auth.api.config.link.command.LinkCommandPathSettings;
 
+import me.mastercapexd.auth.config.discord.DiscordCommandSettings;
+import me.mastercapexd.auth.discord.command.annotation.RenameTo;
 import me.mastercapexd.auth.discord.command.listener.JDACommandListener;
 import me.mastercapexd.auth.hooks.DiscordHook;
 import me.mastercapexd.auth.link.LinkCommandActorWrapper;
 import me.mastercapexd.auth.link.discord.DiscordCommandActorWrapper;
 import me.mastercapexd.auth.link.discord.DiscordLinkType;
 import me.mastercapexd.auth.messenger.commands.MessengerCommandRegistry;
-import me.mastercapexd.auth.discord.command.annotation.RenameTo;
+import me.mastercapexd.auth.messenger.commands.annotation.CommandKey;
 import me.mastercapexd.auth.shared.commands.DiscordLinkCommand;
 import me.mastercapexd.auth.shared.commands.MessengerLinkCommandTemplate;
 import net.dv8tion.jda.api.interactions.commands.OptionType;
 import revxrsal.commands.annotation.dynamic.Annotations;
 import revxrsal.commands.command.CommandActor;
+import revxrsal.commands.command.ExecutableCommand;
 import revxrsal.commands.jda.JDAActor;
 import revxrsal.commands.jda.JDACommandHandler;
 import revxrsal.commands.jda.annotation.OptionData;
@@ -31,7 +35,10 @@ public class DiscordCommandRegistry extends MessengerCommandRegistry {
         COMMAND_HANDLER.registerContextResolver(LinkCommandActorWrapper.class, context -> new DiscordCommandActorWrapper(context.actor()));
         COMMAND_HANDLER.registerAnnotationReplacer(RenameTo.class, (element, parameter) -> Collections.singletonList(
                 Annotations.create(OptionData.class, "value", OptionType.valueOf(parameter.type()), "name", parameter.value())));
-        COMMAND_HANDLER.registerSlashCommandMapper(new DiscordOptionMapper());
+
+        DiscordCommandParameterMapper parameterMapper = new DiscordCommandParameterMapper();
+        COMMAND_HANDLER.registerSlashCommandMapper(parameterMapper);
+        COMMAND_HANDLER.setParameterNamingStrategy(parameterMapper);
         replaceNativeListener();
 
         registerCommands();
